Moved program options to a separate file
